feat: Update GET route in tickets API to retrieve all tickets
This commit updates the GET route in the tickets API to retrieve all tickets from the database. It connects to the MongoDB database using the `connectDB` function and then uses the `Ticket.find()` method to fetch all the tickets. The retrieved tickets are returned as a JSON response. This change ensures that the API route can fetch and return all the tickets to the client.